+++ demo/app/components/editor-layer/editor-layer.component.ts
@@ -17,7 +17,6 @@ export class EditorLayerComponent implements OnInit {
 	}
 	
 	ngOnInit(): void {
-		// this.startEdit();
 	}
 	
 	startEdit() {